Locate solib files via build-id
Hi,

this is a new functionality that the file itself (*) it found via build-id, if
the regular name-located file does not match inferior build-id or if the
named-located file even does not exist.

I am not completely sure with the SO->SO_NAME initialization.  It cannot be
a regression but not sure how it should behave for the build-id-located files.
Currently the original target filename is kept but the comment there says is
may confuse MI frontends.  OTOH putting there the build-id located file will
show its build-id name in SO_NAME which is also wrong.

(*) /usr/lib/libxxx.so, not /usr/lib/debug/usr/lib/libxxx.so.debug.
    Although GDB uses only symbols from that file so in fact the *.debug file
    would be sufficient.  But for example .symtab is sometimes only in the .so
    file and not the .so.debug file so GDB looks up .so as the associated
    .so.debug file gets loaded automatically on top of .so anyway.


Thanks,
Jan


gdb/
2014-03-24  Jan Kratochvil  <jan.kratochvil@redhat.com>

	Locate solib files via build-id.
	* build-id.c (build_id_to_debug_bfd): Rename to ...
	(build_id_to_bfd): ... here.  Add parameter suffix.
	(build_id_to_debug_bfd): New wrapper of build_id_to_bfd.
	(build_id_so_validate): Remove const from the parameter,
	return void.  Handle also NULL SO->ABFD.  Call also build_id_to_bfd.
	* build-id.h (build_id_so_validate): Update declaration.
	* solib.c (solib_map_sections): Call VALIDATE even for NULL SO->ABFD.
	No longer free SO->ABFD on failed VALIDATE.
	(default_solib_validate): Update its const and return value.
	* solib.h (default_solib_validate): Update declaration.
	* solist.h (struct target_so_ops): Update definition of field validate.

gdb/testsuite/
2014-03-24  Jan Kratochvil  <jan.kratochvil@redhat.com>

	Locate solib files via build-id.
	* gdb.base/solib-mismatch.exp (nocrlf, expected_header): Make
	solib_matching_test variables global.
	(solib_locate_test): New proc.
	(locate non-matching shared library)
	(locate non-existing shared library): New tests.